Delete cached interface after stopping
This fixes an issue where the library could use a stale multicast
address.

Steps to reproduce:

1. Connect to a wifi network
2. Start a dnssd browser
3. Stop the dnssd browser
4. Connect to a different wifi network
5. Start a dnssd browser

An error like this is thrown:

Error: setMulticastInterface EADDRNOTAVAIL
    at Socket.setMulticastInterface (dgram.js:566:11)
    at Socket.eval (webpack:///./node_modules/dnssd/lib/NetworkInterface.js?:238:41)
    at Socket.emit (events.js:182:13)
    at startListening (dgram.js:128:10)
    at state.handle.lookup (dgram.js:249:7)
    at process._tickCallback (internal/process/next_tick.js:63:19)

This is because we are still using the IP address from the first
NetworkInterface instance, even though it is not valid for the new
network. Clearing the interface from `activeInterfaces` on stop solves
the problem.